+++ docker/node/docker-entrypoint.sh
@@ -3,17 +3,17 @@ set -e
 
 # first arg is `-f` or `--some-option`
 if [ "${1#-}" != "$1" ]; then
-	set -- yarn "$@"
+	set -- npm "$@"
 fi
 
-if [ "$1" = 'yarn' ]; then
+if [ "$1" = 'npm' ]; then
     # The first time volumes are mounted, the project needs to be recreated
     if [ ! -f package.json ]; then
         # Create a blank project
-		yarn init -yp
-    elif [ "$APP_ENV" != 'prod' ]; then
-        # Always try to reinstall deps when not in prod
-        yarn install
+		npm init -yf
+    else
+        # Always try to reinstall deps
+        npm install
     fi
 fi